% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clustereR}
\alias{clustereR}
\title{clustereR}
\usage{
clustereR(
  ontoNet,
  target,
  method = "leiden",
  filterTerms = c("molecular_function", "intracellular non-membrane-bounded organelle"),
  forceCluster = NULL,
  seed = 42
)
}
\arguments{
\item{ontoNet}{igraph representation of ontology hierarchy from networkeR}

\item{target}{target ontology IDs to cluster}

\item{method}{clustering method (leiden default)}

\item{filterTerms}{terms to not use as a cluster name}

\item{forceCluster}{ontological IDs to force into separate cluster}

\item{seed}{set seed for reproducibility}
}
\value{

}
\description{
clustereR
}
