% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_page.R
\name{clustering_page}
\alias{clustering_page}
\title{Server for the clustering page}
\usage{
clustering_page(input, output, session, descent_data)
}
\arguments{
\item{input}{shiny parameter}

\item{output}{shiny parameter}

\item{session}{shiny parameter}

\item{descent_data}{reactiveValues, contains gene ontology data}
}
\description{
Server for the clustering page
}
