% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_page.R
\name{clustering_page_ui}
\alias{clustering_page_ui}
\title{UI for the clustering page}
\usage{
clustering_page_ui(id)
}
\arguments{
\item{id}{unique id for this module}
}
\description{
UI for the clustering page
}
