% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMS.R
\name{cutText}
\alias{cutText}
\title{Shorten Ontology Terms}
\usage{
cutText(text, cutoff)
}
\arguments{
\item{text}{character vector - ontology terms}

\item{cutoff}{number of characters to cut the terms to}
}
\value{
character vector with shortened ontology terms
}
\description{
Shorten Ontology Terms
}
