% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_entry_page.R
\name{data_entry_page}
\alias{data_entry_page}
\title{Server for the data page}
\usage{
data_entry_page(input, output, session, descent_data)
}
\arguments{
\item{input}{shiny parameter}

\item{output}{shiny parameter}

\item{session}{shiny parameter}

\item{descent_data}{reactiveValues, contains gene ontology data}
}
\description{
Server for the data page
}
