% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMS.R
\name{errorMessage}
\alias{errorMessage}
\title{Create Error Messages}
\usage{
errorMessage(type)
}
\arguments{
\item{type}{character string "empty" or "long", depending on the error message}
}
\value{

}
\description{
Create Error Messages
}
