% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_page.R
\name{exploring_page_ui}
\alias{exploring_page_ui}
\title{UI for the clustering page}
\usage{
exploring_page_ui(id)
}
\arguments{
\item{id}{unique id for this module}
}
\description{
UI for the clustering page
}
