% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{get_test_data}
\alias{get_test_data}
\title{Get test data}
\usage{
get_test_data(size = "big")
}
\arguments{
\item{size}{integer, number of terms to return}
}
\value{
string example_data
}
\description{
Get test data
}
