% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{networkeR}
\alias{networkeR}
\title{networkeR}
\usage{
networkeR(ont = "MF", species = "HSA")
}
\arguments{
\item{ont}{Select ontology type to generate a network. One of c("CC", "BP", "MF","All","Reactome")}

\item{species}{Select species to filter our terms without any genes.}
}
\value{
ontoNet
}
\description{
networkeR
}
