% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_server.R
\name{ontology_descent}
\alias{ontology_descent}
\title{Launch the ontologyDescent shiny server}
\usage{
ontology_descent(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Launch the ontologyDescent shiny server
}
\examples{
\dontrun{
launch_application()
}
}
