% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMS.R
\name{pathwayGraph}
\alias{pathwayGraph}
\title{Plot Enrichment By Pathway}
\usage{
pathwayGraph(
  ontoID = NULL,
  ontoTerm,
  pValue,
  clusterNumber = NULL,
  clusterName,
  enrichmentScore = NULL,
  direction = NULL,
  plotEnrichment = FALSE,
  coordFlip = FALSE,
  themeSet = "bw",
  colorSet = "Brewer",
  lgdPosition = "right",
  nameSize = 7,
  axTxtSize = 7,
  axTitleSize = 9,
  lgTxtSize = 7,
  lgTitleSize = 9,
  fontFam = "sans"
)
}
\arguments{
\item{ontoID}{Optional. A character vector of Gene Ontology IDs.}

\item{ontoTerm}{Optional. A character vector of Terms associated with Gene Ontology IDs.}

\item{pValue}{A numeric vectors of enrichment p Values for each Gene Ontology pathway.}

\item{clusterNumber}{Optional. A numeric vector showing which cluster the Gene Ontologies belong to.}

\item{clusterName}{A character vector showing terms associated with the Cluster IDs.}

\item{enrichmentScore}{Optional. A numeric vector of enrichment scores for each Gene Ontology.}

\item{direction}{Optional. A character vector containing values "Down" or "Up", indicating the direction of pathway enrichment for each Gene Ontology.}

\item{plotEnrichment}{Optional. TRUE or FALSE. Whether to plot the enrichment score instead of p values. Default == FALSE.}

\item{coordFlip}{Optional. TRUE or FALSE. With default == FALSE the plot shows cluster names on y and pValue/enrichmentScore on x axis. When TRUE x and y axes are flipped.}

\item{themeSet}{Optional. A character string to select the visual theme of the plot: one of "bw", "classic", "grey", "minimal" or "dark". Default == "minimal"}

\item{colorSet}{Optional. A character string to select the color palette for the plot: one of  "Brewer", "AAAS", "D3", "Futurama", "IGV", "JAMA", "JCO", "Lancet", "LocusZoom", "NEJM", "NPG", "RickAndMorty", "Simpsons", "StarTrek", "Tron", "UChicago", or "UCSCGB". Default == "IGV"}

\item{lgdPosition}{Optional. A character string setting the legend position: one of "right", "left", "top" or "bottom". Default == "right".}

\item{nameSize}{Optional. A numeric value setting font size for cluster names. Default == 7.}

\item{axTxtSize}{Optional. A numeric value setting font size for the axis text. Defualt == 7.}

\item{axTitleSize}{Optional. A numeric value setting font size for the axis title. Default == 9.}

\item{lgTxtSize}{Optional. A numeric value setting font size for the axis text. Defualt == 7.}

\item{lgTitleSize}{Optional. A numeric value setting font size for the axis text. Defualt == 9.}

\item{fontFam}{Optional. A character string to select the text font family for the plot. One of "serif", "sans" or "mono". Default == "sans".}
}
\value{
A ggplot object.
}
\description{
Plot Enrichment By Pathway
}
