% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_page.R
\name{plotting_page}
\alias{plotting_page}
\title{Server for the plotting page}
\usage{
plotting_page(input, output, session, descent_data)
}
\arguments{
\item{input}{shiny parameter}

\item{output}{shiny parameter}

\item{session}{shiny parameter}

\item{descent_data}{reactiveValues, contains gene ontology data}
}
\description{
Server for the plotting page
}
