% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_entry_page.R
\name{read_input}
\alias{read_input}
\title{Data reader}
\usage{
read_input(x, type)
}
\arguments{
\item{x}{character, data to be read}

\item{type}{character, type of input. Determines which function to read.}
}
\value{

}
\description{
Data reader
}
