% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{relabelleR}
\alias{relabelleR}
\title{Internal Functions}
\usage{
relabelleR(ontoNet, ontoNames, target, filterTerms = c("molecular_function"))
}
\arguments{
\item{ontoNet}{igraph representation of ontology hierarchy from networkeR}

\item{ontoNames}{ontology ID to ontology Terms}

\item{target}{target ontology IDs to cluster}

\item{ontoLength}{ontology ID lengths}
}
\value{

}
\description{
Internal Functions
}
