% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMS.R
\name{reorderData}
\alias{reorderData}
\title{Format Data Frame for Excel Export}
\usage{
reorderData(data)
}
\arguments{
\item{data}{descent_data$inputData}
}
\value{

}
\description{
Format Data Frame for Excel Export
}
