% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_page.R
\name{sorting_page}
\alias{sorting_page}
\title{Server for the sorting page}
\usage{
sorting_page(input, output, session, descent_data)
}
\arguments{
\item{input}{shiny parameter}

\item{output}{shiny parameter}

\item{session}{shiny parameter}

\item{descent_data}{reactiveValues, contains gene ontology data}
}
\description{
Server for the sorting page
}
