% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_entry_page.R
\name{validate_data}
\alias{validate_data}
\title{Validate input data}
\usage{
validate_data(x, type)
}
\arguments{
\item{x}{input data object (must inherit from data.frame)}

\item{type}{character, type of input. Determines which tests to run and phrasing of errors}
}
\value{
string with error messages (length 0 if no errors occured)
}
\description{
Validate input data
}
