% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_entry_page.R
\name{validate_input}
\alias{validate_input}
\title{Validate input}
\usage{
validate_input(x, type)
}
\arguments{
\item{x}{character, input text string (file path or raw data)}

\item{type}{character, type of input. Determines which tests to run and phrasing of errors}
}
\value{
string with error messages (length 0 if no errors occured)
}
\description{
Validate input
}
