% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clustereR}
\alias{clustereR}
\title{Internal Functions}
\usage{
clustereR(
  ontoNet,
  ontoNames,
  ontoLength,
  target,
  method = "edge_betweenness",
  filterTerms = NULL
)
}
\arguments{
\item{ontoNet}{igraph representation of ontology hierarchy from networkeR}

\item{ontoNames}{ontology ID to ontology Terms}

\item{ontoLength}{ontology ID lengths}

\item{target}{target ontology IDs to cluster}
}
\value{

}
\description{
Internal Functions
}
